ensure ChangeProfilePicture calls DeleteFile when file exists and SaveUserPicture throws

+++ src/domain/use-cases/change-profile-picture.ts
@@ -19,7 +19,8 @@ export const setupChangeProfilePicture: Setup = (fileStorage, crypto, userProfil
   try {
     await userProfileRepo.savePicture(userProfile)
   } catch {
-    await fileStorage.delete({ key })
+    if (file !== undefined) await fileStorage.delete({ key })
+    throw new Error()
   }
   return userProfile
 }
